Move legcay motion code around before reintegration
