+++ Ryujinx/Input/Motion/CemuHooks/Client.cs
@@ -3,6 +3,7 @@ using Ryujinx.Common;
 using Ryujinx.Common.Configuration.Hid;
 using Ryujinx.Common.Logging;
 using Ryujinx.Configuration;
+using Ryujinx.Input.Motion.CemuHooks.Protocol;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -11,7 +12,7 @@ using System.Net.Sockets;
 using System.Numerics;
 using System.Threading.Tasks;
 
-namespace Ryujinx.Modules.Motion
+namespace Ryujinx.Input.Motion.CemuHooks
 {
     public class Client : IDisposable
     {
