+++ Ryujinx/Input/MotionInput.cs
@@ -1,7 +1,8 @@
-﻿using System;
+﻿using Ryujinx.Input.Motion;
+using System;
 using System.Numerics;
 
-namespace Ryujinx.Modules.Motion
+namespace Ryujinx.Input
 {
     public class MotionInput
     {
